% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{all_events}
\alias{all_events}
\title{Download all raw event data.}
\usage{
all_events(season)
}
\arguments{
\item{season}{number season to get event data for}
}
\value{
A list of data.frames with the raw event data.
}
\description{
Download all raw event data.
}
