% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{all_players_for_gameday}
\alias{all_players_for_gameday}
\title{Get the list of all players as of a specific Season and Gameday}
\usage{
all_players_for_gameday(season, day)
}
\arguments{
\item{season}{number Season to query}

\item{day}{number Day to query}
}
\value{
data.frame of player information including current team id/nickname
}
\description{
Get the list of all players as of a specific Season and Gameday
}
