% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{count_by_type}
\alias{count_by_type}
\title{Calculate the number of events for each batter or pitcher.}
\usage{
count_by_type(eventType, batterId = NULL, pitcherId = NULL)
}
\arguments{
\item{eventType}{string The type of event to count.}

\item{batterId}{string The ID of the batter(s) by which to filter.}

\item{pitcherId}{string The ID of the pitcher(s) by which to filter.}
}
\value{
The number of events the batter was involved in with the specified type.
}
\description{
Calculate the number of events for each batter or pitcher.
}
