% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{current_roster}
\alias{current_roster}
\title{Get the current roster for a given team}
\usage{
current_roster(teamId = NULL, slug = NULL)
}
\arguments{
\item{teamId}{string The ID of the team (takes precedence if slug is given)}

\item{slug}{string The slug of the team}
}
\value{
data.frame of roster information
}
\description{
Get the current roster for a given team
}
