% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{events}
\alias{events}
\title{Query for game events.}
\usage{
events(
  playerId = NULL,
  gameID = NULL,
  pitcherId = NULL,
  batterId = NULL,
  type = NULL,
  outcomes = TRUE,
  baseRunners = TRUE,
  sortBy = NULL,
  sortDirection = "ASC"
)
}
\arguments{
\item{playerId}{string The ID of a player that must be the batter or the pitcher in each event.}

\item{pitcherId}{string The ID of the pitcher that must be in each event.}

\item{batterId}{string The ID of the batter that must be in each event.}

\item{type}{string The type of event by which to filter.}

\item{outcomes}{boolean Include child player event records.}

\item{baseRunners}{boolean Include child base runner records.}

\item{sortBy}{string The field by which to sort. Most text and numeric columns are supported.}

\item{sortDirection}{string The direction by which to sort. Must be one of ASC, DESC.}

\item{gameId}{string The ID of the game by which to filter results.}
}
\value{
The events that match the query and the number of matching events.
}
\description{
Query for game events.
}
