% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Gets the data from the API server}
\usage{
get_request(path, query = NULL)
}
\arguments{
\item{path}{the endpoint being queried}

\item{query}{additional parameters to the query.}
}
\value{
data from the server, as a data.frame. Or a HTTP error.
}
\description{
Gets the data from the API server
}
