% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{player_ids_by_name}
\alias{player_ids_by_name}
\title{Get all player IDs matching a given player name}
\usage{
player_ids_by_name(name, current = TRUE)
}
\arguments{
\item{name}{string The name of the player}

\item{current}{boolean If true, only players currently using this name will be returned}
}
\value{
data.frame of objects with player IDs matching that name, including the name and timestamps for when that name started and ended being used for that player
}
\description{
Get all player IDs matching a given player name
}
