% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{player_info}
\alias{player_info}
\title{Get extended info for a given player - their name, attributes, ratings, and stars}
\usage{
player_info(playerId = NULL, name = NULL, slug = NULL, all = FALSE)
}
\arguments{
\item{playerId}{string The player ID of the player (takes precedence if other params are specified)}

\item{name}{string The name of the player (takes precedence if slug is specified)}

\item{slug}{string The url_slug of the player}

\item{all}{boolean (default:false) If true, all historical info for the player will be returned, rather than just the current info}
}
\value{
data.frame of objects
}
\description{
Get extended info for a given player - their name, attributes, ratings, and stars
}
