% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{season_leaders}
\alias{season_leaders}
\title{Get the season leaders for a given category and stat}
\usage{
season_leaders(
  season,
  category = NULL,
  stat = NULL,
  order = "ASC",
  limit = NULL
)
}
\arguments{
\item{season}{number Season to query}

\item{category}{string Stat category (batting, pitching, running, or fielding)}

\item{stat}{string Stat to get the leaders for (as returned from /playerStats)}

\item{order}{string Ordering to use for ranking (ASC or DESC)}

\item{limit}{number Number of leaders to return}
}
\value{
data.frame of leaders
}
\description{
Get the season leaders for a given category and stat
}
