% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{tagged_players}
\alias{tagged_players}
\title{Get the list of all players with a modification tag}
\usage{
tagged_players()
}
\value{
data.frame of player information including current team id/nickname
}
\description{
Get the list of all players with a modification tag
}
