% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{all_players}
\alias{all_players}
\title{Get the list of all current players, optionally including players in the Shadows}
\usage{
all_players(includeShadows = FALSE)
}
\arguments{
\item{includeShadows}{boolean whether to include players in the Shadows}
}
\value{
list of players
}
\description{
Get the list of all current players, optionally including players in the Shadows
}
